fixing indexing
